Frontend/api수정임시 (#115)
* Frontend/102-card test

* feat/102-피드페이지스타일링

* chore: fix Url setting

* fix: synth.js 다시 수정 (#107)

* feat: frontend API 수정

* chore: synth.js 수정 코드 재반영

---------

+++ backend/letsnote/src/main/java/com/geeks/letsnote/domain/studio/workSpace/dao/NoteRepository.java
@@ -17,7 +17,7 @@ public interface NoteRepository extends JpaRepository<Note, Long> {
 
     List<Note> findAllBySpaceInstrument(String mapId);
 
-    Optional<Note> findBySpaceInstrumentAndNoteXAndNoteY(String mapId, Integer noteX, Integer noteY);
+    Optional<Note> findBySpaceInstrumentAndNoteXAndNoteY(String mapId, Long noteX, Long noteY);
 
     @Transactional
     @Modifying
